+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## In development:
+- Scale & crop backgrounds to virtual, custom display dimensions
+
 ## v1.5.1 (2019-10-16)
 - scale & crop feature not working on X11 bug 
 [fixed](https://github.com/nwg-piotr/azote/commit/077806e6f72a84fdf768ac6c64d4f081a78fb579)